rename beb filenames to eventmesh
