+++ components/event-publisher-proxy/pkg/commander/eventmesh/eventmesh.go
@@ -1,4 +1,4 @@
-package beb
+package eventmesh
 
 import (
 	"context"
